[8] Move test to better location
